Add some bind mounts to store mailinglist data.
Unfortunately, it's a little more complicated to get our containers
running under separate uids than our host. User namespaces would help
but then there is the tangle of filesystem permissions for data.

Filesystem UID shifting would be great to have to fix this but in the
meantime, punt on the issue.

In the future, we could use Podman with a user namespace plus some
hard-coded uids for each of our containers to stop running as a
privileged host user.